+++ tests/integration-tests/tests/storage/test_raid.py
@@ -22,7 +22,7 @@ from tests.storage.storage_common import verify_directory_correctly_shared
 
 @pytest.mark.regions(["ap-south-1", "us-gov-east-1"])
 @pytest.mark.instances(["c4.xlarge", "c5.xlarge"])
-@pytest.mark.schedulers(["sge", "awsbatch"])
+@pytest.mark.schedulers(["slurm", "awsbatch"])
 @pytest.mark.usefixtures("region", "os", "instance")
 def test_raid_performance_mode(scheduler, pcluster_config_reader, clusters_factory):
     cluster_config = pcluster_config_reader()